% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.summary.mediation.R
\name{bal.summary.mediation}
\alias{bal.summary.mediation}
\title{Compute the balance summary for mediation object.}
\usage{
\method{bal.summary}{mediation}(x, ...)
}
\description{
Compute the balance summary for mediation object.
}
