% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.table.mediation.R
\name{bal.table.mediation}
\alias{bal.table.mediation}
\title{Compute the balance table for mediation object.}
\usage{
\method{bal.table}{mediation}(x, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{object}{A \code{mediation} object}
}
\description{
Compute the balance table for mediation object.
}
