% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal_table.mediation.R
\name{bal_table.mediation}
\alias{bal_table.mediation}
\title{Compute the balance table for mediation object.}
\usage{
\method{bal_table}{mediation}(x, ...)
}
\arguments{
\item{x}{A \code{mediation} object}

\item{...}{list, optional
Additional arguments.}
}
\description{
Compute the balance table for mediation object.
}
