% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal_table.R
\name{bal_table}
\alias{bal_table}
\title{Generic for balance table methods}
\usage{
bal_table(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{list, optional
Additional arguments.}
}
\description{
Generic for balance table methods
}
