% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mediation.R
\name{calculate_effects}
\alias{calculate_effects}
\title{Calculate the actual effects}
\usage{
calculate_effects(w_11, w_00, w_10, w_01, y_outcome, sampw = NULL)
}
\arguments{
\item{w_11}{The Y(1, M(1)) weights}

\item{w_00}{The Y(0, M(0)) weights}

\item{w_10}{The Y(1, M(0)) weights}

\item{w_01}{The Y(0, M(1)) weights}

\item{y_outcome}{The Y variable}
}
\description{
Calculate the actual effects
}
