% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_weights}
\alias{calculate_weights}
\title{Calculate the weights, given propensity scores}
\usage{
calculate_weights(ps, indicator, use_opposite = FALSE)
}
\description{
Calculate the weights, given propensity scores
}
