% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mediation.R
\name{check_equal_wts_stopping}
\alias{check_equal_wts_stopping}
\title{A little helper function that raises an error if the
user provides weights, and they are not equal to the
number of stopping methods}
\usage{
check_equal_wts_stopping(
  weights,
  stopping_methods,
  weights_name = "total_effects_wts"
)
}
\arguments{
\item{weights}{numeric
The weights}

\item{stopping_methods}{The stopping method or methods.}

\item{weights_name}{The name of the weights we are checking.
Default : 'total_effects_wts'}
}
\description{
Call this in the \code{weighted_mediation()} function.
}
