% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mediation.R
\name{check_missing}
\alias{check_missing}
\title{Check vector for NA or NAN values.}
\usage{
check_missing(x)
}
\arguments{
\item{x}{numeric
The data set to check for NA or NAN values.}
}
\description{
\code{check_missing} rasises and error if the data contains.
NA or NAN values.
}
