% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.effects.mediation.R
\name{desc.effects.mediation}
\alias{desc.effects.mediation}
\title{Describe the effects from a mediation object}
\usage{
\method{desc.effects}{mediation}(x, y_outcome = NULL)
}
\arguments{
\item{x}{A mediation object}

\item{y_outcome}{The outcome; if \code{NULL},
then Y must have been provided to the
original mediation function.}
}
\description{
Describe the effects, and calculate standard
errors and confidence intervals from a mediation object
}
