% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.effects.R
\name{desc.effects}
\alias{desc.effects}
\title{Describe the effects}
\usage{
desc.effects(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{list, optional
Additional arguments.}
}
\description{
Describe the effects, and calculate standard
errors and confidence intervals
}
