% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_effects.R
\name{desc_effects}
\alias{desc_effects}
\title{Describe the effects}
\usage{
desc_effects(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{list, optional
Additional arguments.}
}
\description{
Describe the effects, and calculate standard
errors and confidence intervals
}
