% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayPlots.R
\name{displayPlots}
\alias{displayPlots}
\title{Display plots}
\usage{
displayPlots(ptList, figureRows, singlePlot, multiPage, bxpt = FALSE)
}
\arguments{
\item{ptList}{A list of plots to display.}

\item{figureRows}{The number of rows in the figure.}

\item{singlePlot}{An integer indicating the index of the plot to display.}

\item{multiPage}{Whether to display plots on multiple pages.}

\item{bxpt}{Whether to display boxplots. Default: \code{FALSE}.}
}
\description{
Display plots
}
