% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.weights.unstab.R
\name{get.weights.unstab}
\alias{get.weights.unstab}
\title{Extract unstabilized propensity score weights for `iptw` fits}
\usage{
get.weights.unstab(x, stop.method = NULL, withSampW = TRUE)
}
\arguments{
\item{x}{An `iptw` or `mniptw` object.}

\item{stop.method}{The twop method used for the fit of interest.}

\item{withSampW}{Returns weights with sample weights multiplied in, if they were 
provided in the original `iptw` call. Default: `TRUE`.}
}
\value{
Returns a data.frame of weights.
}
\description{
Extracts propensity score weights from an `iptw` or `mniptw`  object.
}
\details{
Weights are the reciprocal of the product of the probability of receiving
the treatment received.
}
\seealso{
[iptw]
}
