% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_mediation.R
\name{is_subset}
\alias{is_subset}
\title{Check whether the variables in d1 are a subset
of the variables in d2. The data sets must be of
equal length.}
\usage{
is_subset(d1, d2)
}
\description{
Check whether the variables in d1 are a subset
of the variables in d2. The data sets must be of
equal length.
}
