% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_mediation.R
\name{mediation_analysis}
\alias{mediation_analysis}
\title{Outcome mediation with \link{gbm} or \link{xgboost}.}
\usage{
mediation_analysis(a_treatment, m_mediator, x_covariates, y_outcome,
  boost_n_trees = 10000, boost_n_folds = 5, boost_maxdepth = 3,
  boost_eta = 0.005, boost_bag_fraction = 1,
  boost_train_fraction = 0.9, boost_subsample = 0.5,
  boost_early_stopping_rounds = 5, booster = "gbm", verbose = FALSE,
  nsims = 100, ...)
}
\arguments{
\item{a_treatment}{integer
The treatement variable, a, which must be
dichotomous.}

\item{m_mediator}{numeric
The mediator variable, m.}

\item{x_covariates}{numeric
The covariates, x}

\item{y_outcome}{numeric
The outcome variable, y. If this is not provided, then
no outcomes will be calculated
(Default : \code{NULL}).}

\item{boost_n_trees}{integer
Number of iterations passed on to \link{gbm} or \link{xgboost}
(Default : \code{10000}).}

\item{boost_n_folds}{integer
Number of cross-validation folds, passed only to \link{xgboost}
(Default : \code{5}).}

\item{boost_maxdepth}{integer
The maximum interaction depth passed on to \link{gbm} or \link{xgboost}
(Default : \code{3}).}

\item{boost_eta}{numerc
Shrinkage passed on to \link{gbm} or \link{xgboost}
(Default : \code{0.005}).}

\item{boost_bag_fraction}{numeric
Bag fraction passed on to \link{gbm}
(Default : \code{1.0}).}

\item{boost_train_fraction}{numeric
Bag training fraction passed only to \link{gbm}
(Default : \code{0.9}).}

\item{boost_subsample}{numeric
The subsample fraction, passed to \link{xgboost}
(Default : \code{0.5})}

\item{boost_early_stopping_rounds}{integer
The number of rounds of no change after which to
to stop early, passsed to \link{xgboost}
(Default : \code{5})}

\item{booster}{character
Either \code{gbm} or \code{xgboost}.
(Default : \code{'gbm'}).}

\item{verbose}{If \code{TRUE}, lots of information will be printed to monitor
the progress of the fitting.
(Default: \code{TRUE}.)}

\item{nsims}{integer
The number of simulations to perform.
(Default : \code{100})}
}
\description{
Outcome mediation with \link{gbm} or \link{xgboost}.
}
