% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mediation.R
\name{plot.mediation}
\alias{plot.mediation}
\title{Plot the \code{mediation} object.}
\usage{
\method{plot}{mediation}(
  x,
  plots = "optimize",
  subset = NULL,
  color = TRUE,
  model_subset = NULL,
  ...
)
}
\arguments{
\item{x}{weighted_mediation object}

\item{plots}{An indicator of which type of plot is desired. The options are
\itemize{
\item \verb{"optimize" or 1} A plot of the balance criteria as a function of the GBM
iteration.
\item \verb{"boxplot" or 2} Boxplots of the propensity scores for the treatment and
control cases
\item \verb{"es" or 3} Plots of the standardized effect size of the pre-treatment
variables before and after reweighing
\item \verb{"t" or 4} Plots of the p-values from t-statistics comparing means of
treated and control subjects for pretreatment variables, before and after
weighting.
\item \verb{"ks" or 5} Plots of the p-values from Kolmogorov-Smirnov statistics
comparing distributions of pretreatment variables of treated and control
subjects, before and after weighting.
}}

\item{subset}{Used to restrict which of the \code{stop.method}s will be used
in the figure. For example \code{subset = c(1,3)} would indicate that the
first and third \code{stop.method}s (in alphabetical order of those specified
in the original call to the mediation function) should be included in the
figure.}

\item{color}{If \code{color = FALSE}, figures will be gray scale. Default: \code{TRUE}.}

\item{model_subset}{integer
Choose either model A or model M only.}

\item{...}{Additional arguments.}
}
\description{
Plot the \code{mediation} object.
}
