% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ps.R
\name{plot.ps}
\alias{plot.ps}
\title{Plots for `ps` objects}
\usage{
\method{plot}{ps}(x, plots = "optimize", subset = NULL, color = TRUE,
  ...)
}
\arguments{
\item{x}{A `ps` object.}

\item{plots}{An indicator of which type of plot is desired. The options are
* `"optimize" or 1` A plot of the balance criteria as a function of the GBM 
  iteration.
* `"boxplot" or 2` Boxplots of the propensity scores for the treatment and 
  control cases
* `"es" or 3` Plots of the standardized effect size of the pre-treatment 
  variables before and after reweighing
* `"t" or 4` Plots of the p-values from t-statistics comparing means of 
  treated and control subjects for pretreatment variables, before and after 
  weighting.
* `"ks" or 5` Plots of the p-values from Kolmogorov-Smirnov statistics 
  comparing distributions of pretreatment variables of treated and control 
  subjects, before and after weighting.}

\item{subset}{If multiple `stop.method` rules were used in the `ps()` call,
`subset` restricts the plots of a subset of the stopping rules that were
employed. This argument expects a subset of the integers from 1 to k,
if k `stop.method`s were used.}

\item{color}{If `color = FALSE`, figures will be gray scale. Default: `TRUE`.}

\item{...}{Additional arguments.}
}
\description{
This function produces a collection of diagnostic plots for `ps` objects.
}
\details{
This function produces lattice-style graphics of diagnostic plots.
}
\references{
Dan McCaffrey, G. Ridgeway, Andrew Morral (2004). "Propensity
  Score Estimation with Boosted Regression for Evaluating Adolescent
  Substance Abuse Treatment", *Psychological Methods* 9(4):403-425.
}
\seealso{
[ps]
}
\keyword{models,}
\keyword{multivariate}
