\name{print.dxwts}
\alias{print.dxwts}

\title{ Print a diagnosis of the weights }
\description{
Prints a diagnosis of the weights. Extracts \code{summary.tab} from the 
\code{\link[twang]{dxwts}} object
}
\usage{
print.dxwts(x, ...)
}

\arguments{
  \item{x}{ a \code{\link[twang]{dxwts}} object}
  \item{\dots}{further arguments passed to or from other methods}
}

\value{
See \code{\link[twang]{ps}} for a description of the components of the table
}
\keyword{print}
