% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mediation.R
\name{print.mediation}
\alias{print.mediation}
\title{Default print statement for \code{mediation} class}
\usage{
\method{print}{mediation}(object, ...)
}
\arguments{
\item{object}{A \code{mediation} object}

\item{...}{Additional arguments.}
}
\description{
Default print statement for \code{mediation} class
}
