% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mniptw.R
\name{print.mniptw}
\alias{print.mniptw}
\title{Default print statement for `mniptw` class}
\usage{
\method{print}{mniptw}(x, ...)
}
\arguments{
\item{x}{A `mniptw` object}

\item{...}{Additional arguments.}
}
\description{
Default print statement for `mniptw` class
}
