% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ps.R
\name{print.ps}
\alias{print.ps}
\title{Default print statement for `ps` class}
\usage{
\method{print}{ps}(x, ...)
}
\arguments{
\item{x}{An `ps` object}

\item{...}{Additional arguments.}
}
\description{
Default print statement for `ps` class
}
