% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.iptw.R
\name{print.summary.iptw}
\alias{print.summary.iptw}
\title{Produces a summary table for `iptw` object}
\usage{
\method{print}{summary.iptw}(x, ...)
}
\arguments{
\item{x}{An `iptw` object}

\item{...}{Additional arguments.}
}
\description{
Produces a summary table for `iptw` object
}
