% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mniptw.R
\name{print.summary.mniptw}
\alias{print.summary.mniptw}
\title{Produces a summary table for \code{mniptw} object}
\usage{
\method{print}{summary.mniptw}(x, ...)
}
\arguments{
\item{x}{An \code{mniptw} object}

\item{...}{Additional arguments.}
}
\description{
Produces a summary table for \code{mniptw} object
}
