% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.mnps.R
\name{print.summary.mnps}
\alias{print.summary.mnps}
\title{Produces a summary table for \code{mnps} object}
\usage{
\method{print}{summary.mnps}(x, ...)
}
\arguments{
\item{x}{An \code{mnps} object}

\item{...}{Additional arguments.}
}
\description{
Produces a summary table for \code{mnps} object
}
