% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ps.R
\name{print.summary.ps}
\alias{print.summary.ps}
\title{Produces a summary table for \code{ps} object}
\usage{
\method{print}{summary.ps}(x, ...)
}
\arguments{
\item{x}{An \code{ps} object}

\item{...}{Additional arguments.}
}
\description{
Produces a summary table for \code{ps} object
}
