% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mediation.R
\name{summary.mediation}
\alias{summary.mediation}
\title{Displays a useful description of a \code{mediation} object.}
\usage{
\method{summary}{mediation}(object, ...)
}
\arguments{
\item{object}{is a \code{mediation} object}

\item{...}{Additional arguments.}
}
\description{
Displays a useful description of a \code{mediation} object.
}
