% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mniptw.R
\name{summary.mniptw}
\alias{summary.mniptw}
\title{Summarize a `mniptw` object}
\usage{
\method{summary}{mniptw}(object, ...)
}
\arguments{
\item{object}{An `mniptw` object.}

\item{...}{Additional arguments.}
}
\description{
Summarize a `mniptw` object
}
