% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mnps.R
\name{summary.mnps}
\alias{summary.mnps}
\title{Summarize a \code{mnps} object}
\usage{
\method{summary}{mnps}(object, ...)
}
\arguments{
\item{object}{An \code{mnps} object.}

\item{...}{Additional arguments.}
}
\value{
See \link{mnps} for details on the returned table.
}
\description{
Computes summary information about a stored \code{mnps} object
}
\details{
Compresses the information in the \code{desc} component of the \code{mnps} object
into a short summary table describing the size of the dataset and the quality of
the propensity score weights.
}
\seealso{
\link{mnps}
}
\keyword{models}
