% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ps.R
\name{summary.ps}
\alias{summary.ps}
\title{Summarize a `ps` object}
\usage{
\method{summary}{ps}(object, ...)
}
\arguments{
\item{object}{An `ps` object.}

\item{...}{Additional arguments.}
}
\value{
See [ps] for details on the returned table.
}
\description{
Computes summary information about a stored `ps` object
}
\details{
Compresses the information in the `desc` component of the `ps` object
into a short summary table describing the size of the dataset and the quality of
the propensity score weights.
}
\seealso{
[ps]
}
\keyword{models}
