% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.table.mediation.R
\name{bal.table.mediation}
\alias{bal.table.mediation}
\title{Compute the balance table for mediation object.}
\usage{
\method{bal.table}{mediation}(x, digits = 3, ...)
}
\arguments{
\item{x}{A \code{mediation} object}

\item{...}{Additional arguments.}
}
\description{
Compute the balance table for mediation object.
}
