% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iptw.R
\name{iptw}
\alias{iptw}
\title{Inverse probability of treatment weighting for marginal structural models.}
\usage{
iptw(formula, data, timeInvariant = NULL, cumulative = TRUE,
  timeIndicators = NULL, ID = NULL, priorTreatment = TRUE,
  n.trees = 10000, nrounds = n.trees, interaction.depth = 3,
  max_depth = interaction.depth, shrinkage = 0.01, eta = shrinkage,
  bag.fraction = 1, subsample = bag.fraction, params = NULL,
  perm.test.iters = 0, print.level = 2, verbose = TRUE,
  stop.method = c("es.max"), sampw = NULL, version = "fast",
  ks.exact = NULL, booster = "gbm", tree_method = "hist",
  save.propensities = FALSE, file = NULL, n.keep = 1, n.grid = 25,
  ...)
}
\arguments{
\item{formula}{Either a single formula (long format) or a list with formulas.}

\item{data}{The dataset, includes treatment assignment as well as covariates.}

\item{timeInvariant}{An optional formula (with no left-hand variable) specifying time-invariant
chararacteristics.}

\item{cumulative}{If \code{TRUE}, the time t model includes time-varying characteristics from times 1
through t-1. Default: \code{TRUE}.}

\item{timeIndicators}{For long format fits, a vector of times for each observation.}

\item{ID}{For long format fits, a vector of numeric identifiers for unique analytic units.}

\item{priorTreatment}{For long format fits, includes treatment levels from previous times if \code{TRUE}. This
argument is ignored for wide format fits. Default: \code{TRUE}.}

\item{n.trees}{Number of gbm iterations passed on to \link{gbm}.}

\item{interaction.depth}{A positive integer denoting the tree depth used in
gradient boosting. Only one of \code{interaction.depth} and
\code{max_depth} can be specified. Default: 3.}

\item{shrinkage}{A numeric value between 0 and 1 denoting the learning rate.
See \link{gbm} for more details. Only one of \code{shrinkage}
and \code{eta} can be specified. Default: 0.01.}

\item{bag.fraction}{A numeric value between 0 and 1 denoting the fraction of
the observations randomly selected in each iteration of the gradient
boosting algorithm to propose the next tree. See \link{gbm} for
more details. Only one of \code{bag.fraction} and \code{subsample} can be
specified. Default: 1.0.}

\item{params}{...}

\item{perm.test.iters}{A non-negative integer giving the number of iterations
of the permutation test for the KS statistic. If \code{perm.test.iters=0}
then the function returns an analytic approximation to the p-value. Setting
\code{perm.test.iters=200} will yield precision to within 3\% if the true
p-value is 0.05. Use \code{perm.test.iters=500} to be within 2\%. Default: 0.}

\item{print.level}{The amount of detail to print to the screen. Default: 2.}

\item{verbose}{If \code{TRUE}, lots of information will be printed to monitor the
the progress of the fitting. Default: \code{TRUE}.}

\item{stop.method}{A method or methods of measuring and summarizing balance across pretreatment
variables. Current options are \code{ks.mean}, \code{ks.max}, \code{es.mean}, and \code{es.max}. \code{ks} refers to the
Kolmogorov-Smirnov statistic and es refers to standardized effect size. These are summarized
across the pretreatment variables by either the maximum (\code{.max}) or the mean (\code{.mean}).
Default: \code{c("es.max")}.}

\item{sampw}{Optional sampling weights.}

\item{version}{Legacy?}

\item{ks.exact}{\code{NULL} or a logical indicating whether the
Kolmogorov-Smirnov p-value should be based on an approximation of exact
distribution from an unweighted two-sample Kolmogorov-Smirnov test. If
\code{NULL}, the approximation based on the exact distribution is computed
if the product of the effective sample sizes is less than 10,000.
Otherwise, an approximation based on the asymptotic distribution is used.
\strong{Warning:} setting \code{ks.exact = TRUE} will add substantial
computation time for larger sample sizes. Default: \code{NULL}.}

\item{booster}{\code{gbm} or \code{xgboost}. Default: \code{gbm}.}

\item{tree_method}{\code{xgboost} param. Default: "hist".}

\item{save.propensities}{Whether to save the propensity scores. Default: \code{FALSE}.}

\item{file}{Optional RDS file path where the \code{ps} object will be saved.}

\item{n.keep}{A numeric variable indicating the algorithm should only
consider every \code{n.keep}-th iteration of the propensity score model and
optimize balance over this set instead of all iterations. Default: 1.}

\item{n.grid}{A numeric variable that sets the grid size for an initial
search of the region most likely to minimize the \code{stop.method}. A
value of \code{n.grid=50} uses a 50 point grid from \code{1:n.trees}. It
finds the minimum, say at grid point 35. It then looks for the actual
minimum between grid points 34 and 36. Default: 25.}

\item{...}{Additional arguments that are passed to ps function.}
}
\value{
Returns an object of class \code{iptw}, a list containing
\itemize{
\item \code{psList} A list of \code{ps} objects with length equal to the number of time periods.
\item \code{estimand} The specified estimand.
\item \code{stop.methods} The stopping rules used to optimize \code{iptw} balance.
\item \code{nFits} The number of \code{ps} objects (i.e., the number of distinct time points).
\item \code{uniqueTimes} The unique times in the specified model.
}
}
\description{
\code{iptw} uses \link{gbm} to estimate propensity scores for sequential treatments.
}
\details{
This function uses generalized boosted models to estimate inverse probability of
treatment weights for sequential treatments.
}
\seealso{
\link{ps}
}
