% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dxwts.R
\name{plot.dxwts}
\alias{plot.dxwts}
\title{Plot \code{dxwts}}
\usage{
\method{plot}{dxwts}(x, plots = "es", ...)
}
\arguments{
\item{x}{An \code{dxwts} object.}

\item{plots}{An indicator of which type of plot is desired. The options are
\itemize{
\item \code{"optimize" or 1} A plot of the balance criteria as a function of the GBM
iteration.
\item \code{"boxplot" or 2} Boxplots of the propensity scores for the treatment and
control cases
\item \code{"es" or 3} Plots of the standardized effect size of the pre-treatment
variables before and after reweighing
\item \code{"t" or 4} Plots of the p-values from t-statistics comparing means of
treated and control subjects for pretreatment variables, before and after
weighting.
\item \code{"ks" or 5} Plots of the p-values from Kolmogorov-Smirnov statistics
comparing distributions of pretreatment variables of treated and control
subjects, before and after weighting.
}}

\item{...}{Additional arguments.}
}
\description{
Plot \code{dxwts}
}
