% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mediation.R
\name{plot.mediation}
\alias{plot.mediation}
\title{Plot the \code{mediation} object.}
\usage{
\method{plot}{mediation}(
  x,
  plots = "optimize",
  subset = NULL,
  color = TRUE,
  model_subset = NULL,
  ...
)
}
\arguments{
\item{x}{weighted_mediation object}

\item{plots}{An indicator of which type of plot is desired. The options are
\itemize{
\item \code{"optimize"} A plot of the balance criteria as a function of the GBM
iteration.
\item \code{"boxplot"} Boxplots of the propensity scores for the treatment and
control cases
\item \code{"es"} Plots of the standardized effect size of the pre-treatment
variables before and after reweighing
\item \code{"density"} Distriubtion plots of NIE1 (distribution of mediator for treatment
sample weighted to match distribution of mediator under control for the population)
and NIE0 (distribution of mediator for control sample weighted to match
distribution of mediator under treatment for the population) for each mediator.
For continuous mediators, distributions are plotted with density curves and
for categorical (factor) mediators, distributions are plotted with barplots.
}}

\item{subset}{Used to restrict which of the \code{stop.method}s will be used
in the figure. For example \code{subset = c(1,3)} would indicate that the
first and third \code{stop.method}s (in alphabetical order of those specified
in the original call to the mediation function) should be included in the
figure. If x$method = \code{logistic} or \code{crossval}, there is no need to subset
as there is only one method used.}

\item{color}{If \code{color = FALSE}, figures will be gray scale. Default: \code{TRUE}.}

\item{model_subset}{integer
Choose either model A or model M only.}

\item{...}{Additional arguments.}
}
\description{
Plot the \code{mediation} object.
}
