% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mnps.R
\name{plot.mnps}
\alias{plot.mnps}
\title{Plots for `mnps` objects}
\usage{
\method{plot}{mnps}(x, plots = "optimize", pairwiseMax = TRUE,
  figureRows = NULL, color = TRUE, subset = NULL,
  treatments = NULL, singlePlot = NULL, multiPage = FALSE,
  time = NULL, print = TRUE, hline = c(0.1, 0.5, 0.8), ...)
}
\arguments{
\item{x}{An `mnps` object.}

\item{plots}{An indicator of which type of plot is desired. The options are
* `"optimize" or 1` A plot of the balance criteria as a function of the GBM 
  iteration.
* `"boxplot" or 2` Boxplots of the propensity scores for the treatment and 
  control cases
* `"es" or 3` Plots of the standardized effect size of the pre-treatment 
  variables before and after reweighing
* `"t" or 4` Plots of the p-values from t-statistics comparing means of 
  treated and control subjects for pretreatment variables, before and after 
  weighting.
* `"ks" or 5` Plots of the p-values from Kolmogorov-Smirnov statistics 
  comparing distributions of pretreatment variables of treated and control 
  subjects, before and after weighting.}

\item{pairwiseMax}{If `FALSE`, the plots for the underlying `ps` fits 
will be returned.  Otherwise, pairwise maxima will be returned.}

\item{figureRows}{The number of rows of figures that should be used.
If left as `NULL`, twang tries to find a reasonable value.}

\item{color}{If `color = FALSE`, figures will be gray scale. Default: `TRUE`.}

\item{subset}{Used to restrict which of the `stop.method`s will be used 
in the figure. For example `subset = c(1,3)` would indicate that the 
first and third `stop.method`s (in alphabetical order of those specified 
in the original call to `mnps`) should be included in the figure.}

\item{treatments}{Only applicable when `pairwiseMax` is `FALSE` and `plots` 3, 4, and 5.
If left at `NULL`, panels for all treatment pairs are created.  If one level of the treatment 
variable is specified, plots comparing that treatment to all others are produced.  If two
levels are specified, a comparison for that single pair is produced.}

\item{singlePlot}{For Plot calls that produce multiple plots, specifying an integer value of 
`singlePlot` will return only the corresponding plot.  E.g., specifying `singlePlot = 2` 
will return the second plot.}

\item{multiPage}{When multiple frames of a figure are produced, `multiPage = TRUE` will print each
frame on a different page. This is intended for situations where the graphical output is being
saved to a file.}

\item{time}{For use with `iptw`.}

\item{print}{If `FALSE`, the figure is returned but not printed. Default: `TRUE`.}

\item{hline}{Arguments passed to `panel.abline`.}

\item{...}{Additional arguments.}
}
\description{
This function produces a collection of diagnostic plots for `mnps` objects.
}
\details{
This function produces lattice-style graphics of diagnostic plots.
}
\references{
Dan McCaffrey, G. Ridgeway, Andrew Morral (2004). "Propensity
  Score Estimation with Boosted Regression for Evaluating Adolescent
  Substance Abuse Treatment", *Psychological Methods* 9(4):403-425.
}
\seealso{
[mnps]
}
\keyword{models,}
\keyword{multivariate}
