% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dxwts.R
\name{print.dxwts}
\alias{print.dxwts}
\title{Default print statement for `dxwts` class}
\usage{
\method{print}{dxwts}(x, ...)
}
\arguments{
\item{x}{A `dxwts` object}

\item{...}{Additional arguments.}
}
\description{
Default print statement for `dxwts` class
}
