\name{print.dxwts}
\alias{print.dxwts}

\title{ Print a diagnosis of the weights }
\description{
Prints a diagnosis of the weights. Extracts \code{summary.tab} from the 
\code{\link{dx.wts}} object
}

\usage{
\method{print}{dxwts}(x, ...)
}

\arguments{
  \item{x}{ a \code{\link{dx.wts}} object}
  \item{\dots}{further arguments passed to or from other methods}
}

\value{
See \code{\link{ps}} for a description of the components of the table
}
\keyword{print}
