% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.iptw.R
\name{print.iptw}
\alias{print.iptw}
\title{Default print statement for `iptw` class}
\usage{
\method{print}{iptw}(x, ...)
}
\arguments{
\item{x}{A `iptw` object}

\item{...}{Additional arguments.}
}
\description{
Default print statement for `iptw` class
}
