% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mnps.R
\name{print.mnps}
\alias{print.mnps}
\title{Default print statement for \code{mnps} class}
\usage{
\method{print}{mnps}(x, ...)
}
\arguments{
\item{x}{A \code{mnps} object}

\item{...}{Additional arguments.}
}
\description{
Default print statement for \code{mnps} class
}
