% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{ps}
\alias{ps}
\title{Gradient boosted propensity score estimation}
\usage{
ps(formula = formula(data), data, n.trees = 10000, nrounds = n.trees,
  interaction.depth = 3, max_depth = interaction.depth,
  shrinkage = 0.01, eta = shrinkage, bag.fraction = 1,
  subsample = bag.fraction, params = NULL, perm.test.iters = 0,
  print.level = 2, verbose = TRUE, estimand = "ATE",
  stop.method = c("ks.mean", "es.mean"), sampw = NULL,
  multinom = FALSE, version = "fast", ks.exact = NULL,
  booster = "gbm", tree_method = "hist", n.keep = 1, n.grid = 25,
  ...)
}
\arguments{
\item{formula}{an object of class "\code{\link{formula}}": a symbolic
description of the propensity score model to be fit with the treatment
indicator on the left side of the formula and the potential confounding
variables on the right side.}

\item{data}{a dataset that includes the treatment indicator as well as the
potential confounding variables.}

\item{n.trees}{a positive integer denoting the maximum number of iterations
to be used in the gradient boosting algorithm. Only one of \code{n.trees}
and \code{nrounds} can be specified. Default: 10000}

\item{interaction.depth}{a positive integer denoting the tree depth used in
gradient boosting. Only one of \code{interaction.depth} and
\code{max_depth} can be specified. Default: 3}

\item{shrinkage}{a numeric value between 0 and 1 denoting the learning rate.
See \code{\link[gbm]{gbm}} for more details. Only one of \code{shrinkage}
and \code{eta} can be specified. Default: 0.01}

\item{bag.fraction}{a numeric value between 0 and 1 denoting the fraction of
the observations randomly selected in each iteration of the gradient
boosting algorithm to propose the next tree. See \code{\link[gbm]{gbm}} for
more details. Only one of \code{bag.fraction} and \code{subsample} can be
specified. Default: 1}

\item{perm.test.iters}{a non-negative integer giving the number of iterations
of the permutation test for the KS statistic. If \code{perm.test.iters=0}
then the function returns an analytic approximation to the p-value. Setting
\code{perm.test.iters=200} will yield precision to within 3\% if the true
p-value is 0.05. Use \code{perm.test.iters=500} to be within 2\%. Default:
0}

\item{print.level}{the amount of detail to print to the screen}

\item{verbose}{if TRUE, lots of information will be printed to monitor the
the progress of the fitting}

\item{estimand}{"ATE" (average treatment effect) or "ATT" (average treatment
effect on the treated) : the causal effect of interest. ATE estimates the
change in the outcome if the treatment were applied to the entire
population versus if the control were applied to the entire population. ATT
estimates the analogous effect, averaging only over the treated population.
Default: "ATE"}

\item{stop.method}{a character vector of the method or methods to be used to
optimize the balance of pretreatment variables. Current options are
\code{ks.mean}, \code{ks.max}, \code{es.mean}, and \code{es.max}. \code{ks}
refers to the Kolmogorov-Smirnov statistic and \code{es} refers to
standardized effect size. These are summarized across the pretreatment
variables by either the maximum (\code{.max}) or the mean (\code{.mean}).}

\item{sampw}{Optional sampling weights.}

\item{version}{Legacy?}

\item{ks.exact}{\code{NULL} or a logical indicating whether the
Kolmogorov-Smirnov p-value should be based on an approximation of exact
distribution from an unweighted two-sample Kolmogorov-Smirnov test. If
\code{NULL}, the approximation based on the exact distribution is computed
if the product of the effective sample sizes is less than 10,000.
Otherwise, an approximation based on the asymptotic distribution is used.
\strong{Warning:} setting \code{ks.exact = TRUE} will add substantial
computation time for larger sample sizes.}

\item{booster}{gbm or xgboost}

\item{tree_method}{xgboost param}

\item{n.keep}{a numeric variable indicating the algorithm should only
consider every \code{n.keep}-th iteration of the propensity score model and
optimize balance over this set instead of all iterations.}

\item{n.grid}{a numeric variable that sets the grid size for an initial
search of the region most likely to minimize the \code{stop.method}. A
value of \code{n.grid=50} uses a 50 point grid from \code{1:n.trees}. It
finds the minimum, say at grid point 35. It then looks for the actual
minimum between grid points 34 and 36.}
}
\description{
\code{ps} calculates propensity scores using gradient boosted logistic
regression and diagnoses the resulting propensity scores using a variety of
methods
}
\details{
For user more comfortable with the options of \code{\link[xgboost]{xgboost}},
the options for \code{ps} controlling the behavior of the gradient boosting
algorithm can be specified using the \code{\link[xgboost]{xgboost}} naming
scheme. This includes \code{nrounds}, \code{max_depth}, \code{eta}, and
\code{subsample}. In addition, the list of parameters passed to
\code{\link[xgboost]{xgboost}} can be specified with \code{params}.

Note that unlike earlier versions of \code{twang}, the plotting functions are
no longer included in the \code{ps} function. See  \code{\link{plot}} for
details of the plots.
}
\references{
Dan McCaffrey, G. Ridgeway, Andrew Morral (2004). "Propensity
  Score Estimation with Boosted Regression for Evaluating Adolescent
  Substance Abuse Treatment", \emph{Psychological Methods} 9(4):403-425.
}
