% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop.methods.R
\docType{data}
\name{stop.methods}
\alias{stop.methods}
\title{Stop methods (e.g. "es.mean", "ks.mean", etc.) object, used only for backward compatibility}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 1 rows and 6 columns.
}
\usage{
stop.methods
}
\description{
In older versions of twang, the \code{ps} function specified the \code{stop.method} in a different
manner. This \code{stop.methods} object is used to ensure backward compatibility; new twang
users should not make use of it.
}
\details{
This is merely a vector with the names of the stopping rules.
}
\keyword{datasets}
