\name{stop.methods}
\alias{stop.methods}

\title{Object only used for backward compatibility}
\description{
  In older versions of twang, the \code{ps} function specified the \code{stop.method} in a different manner.  This \code{stop.methods} object is used to ensure backward compatibility; new twang users should not make use of it.
}
\details{
	
	This is merely a vector with the names of the stopping rules.  
}


\seealso{ \code{\link{ps}} }

\keyword{ utilities }
