% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mnps.R
\name{summary.mnps}
\alias{summary.mnps}
\title{Summarize a `mnps` object}
\usage{
\method{summary}{mnps}(object, ...)
}
\arguments{
\item{object}{An `mnps` object.}

\item{...}{Additional arguments.}
}
\value{
See [mnps] for details on the returned table.
}
\description{
Computes summary information about a stored `mnps` object
}
\details{
Compresses the information in the `desc` component of the `mnps` object
into a short summary table describing the size of the dataset and the quality of
the propensity score weights.
}
\seealso{
[mnps]
}
\keyword{models}
