% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twang.R
\docType{package}
\name{twang-package}
\alias{twang}
\alias{twang-package}
\title{twang: Toolkit for Weighting and Analysis of Nonequivalent Groups}
\description{
Provides functions for propensity score
        estimating and weighting, nonresponse weighting, and diagnosis
        of the weights.
}
\keyword{propensity}
\keyword{score}
