% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mediation.R
\name{weighted_mediation}
\alias{weighted_mediation}
\title{Weighted mediation analysis.}
\usage{
weighted_mediation(a_treatment, m_mediator, x_covariates = NULL,
  x_covariates_mediator = NULL, y_outcome = NULL, ps_n.trees = 10000,
  ps_interaction.depth = 3, ps_shrinkage = 0.005,
  ps_bag.fraction = 1, ps_perm.test.iters = 0, ps_iterlim = 1000,
  ps_stop.method = c("ks.mean", "ks.max"), ps_sampw = NULL,
  ps_obj = NULL, ps_weights = NULL, ...)
}
\arguments{
\item{a_treatment}{integer
The treatement variable, a, which must be
dichotomous.}

\item{m_mediator}{numeric
The mediator variable, m.}

\item{x_covariates}{numeric
The covariates, x
(Default : \code{NULL}).}

\item{x_covariates_mediator}{numeric
The covariates, x, if different for the mediator model
(Default : \code{NULL}).}

\item{y_outcome}{numeric
The outcome variable, y. If this is not provided, then
no outcomes will be calculated
(Default : \code{NULL}).}

\item{ps_n.trees}{integer
Number of gbm iterations passed on to \code{gbm}
(Default : \code{10000}).}

\item{ps_interaction.depth}{integer
Interaction.depth passed on to \code{gbm}
(Default : \code{3}).}

\item{ps_shrinkage}{numerc
Shrinkage passed on to \code{gbm}
(Default : \code{0.005}).}

\item{ps_bag.fraction}{numerc
Bag fraction passed on to \code{gbm}
(Default : \code{1.0}).}

\item{ps_perm.test.iters}{integer
A non-negative integer giving the number of iterations
of the permutation test, passed to \code{gbm}
(Default : \code{0}).}

\item{ps_iterlim}{integer
Maximum number of iterations for the direct optimization
(Default : \code{1000}).}

\item{ps_stop.method}{integer
A method or methods of measuring and summarizing balance
across pretreament variables. Current options are \code{ks.mean},
\code{ks.max}, \code{es.mean}, and \code{es.max}
(Default : \code{c("ks.mean", "ks.max")}).}

\item{ps_sampw}{numeric
Optional sampling weights
(Default : \code{NULL}).}

\item{ps_obj}{ps object
Optional \code{ps} object. If this is provided
it will be used instead of fitting a model
with \code{x_covariates}.
(Default : \code{NULL}).}

\item{ps_weights}{data frame
Optional data.frame object. If this is provided
it will be used instead of fitting a model
with \code{x_covariates}.
(Default : \code{NULL}).}

\item{...}{list
Additional arguments passed to \link[twang:ps]{twang::ps}.}
}
\value{
mediation object
The \code{mediation} object includes the following:
\itemize{
\item \code{model_a} The model A \code{ps()} results.
\item \code{model_m} The model M \code{ps()} results.
\item \code{model_a_weights} The model A weights.
\item \code{model_m_weights} The model M weights.
\item \code{natural_indirect_weights} The natural indirect weights.
\item \code{data} The data set used to compute model M.
\item \code{stopping_methods} The stopping methods passed to \code{stop.method}.
\item \code{datestamp} The date when the analysis was run.
\item For each \code{stop.method}, a list with the following:
\itemize{
\item \code{overall_effect} The overall effect.
\item \code{natural_direct_effect} The natural direct effect.
\item \code{natural_indirect_effect} The natural indirect effect.
\item \code{expected_treatment0_mediator0} E(Y(0, M(0)))
\item \code{expected_treatment1_mediator1} E(Y(1, M(1)))
\item \code{expected_treatment1_mediator0} E(Y(0, M(1)))
}
}
}
\description{
Estimate causal mediation mechanism of a treatment under
using propensity score weighting.
}
\details{
For users comfortable with \link{ps}, any options prefaced with
\code{ps_} are passed directly to the \code{ps()} function.
}
\seealso{
\link{ps}
}
\keyword{models,}
\keyword{multivariate}
